Create ADnD_2nd_Edition.html
This is my version of an Advanced Dungeons and Dragons 2nd Edition character sheet, using Actoba's slick character sheet with his permission, hacked and pieced together. I am NOT a coder, so if you see something that looks horrible but it works, either deal with it, or fix it!

all of the spacing you see at the top is due to me NOT understanding how tabbing works/is done.